+++ app.yaml
@@ -1,10 +1,14 @@
-runtime: python37
+runtime: custom
+env: flex
+
+#entrypoint: gunicorn -b :$PORT main:app
 
 handlers:
   # This configures Google App Engine to serve the files in the app's static
   # directory.
 - url: /static
-  static_dir: /app/static
+  static_dir: static
+
 
   # This handler routes all requests not caught above to your main app. It is
   # required when static routes are defined, but can be omitted (along with
