--- app/__init__.py
@@ -1,3 +1,0 @@
-from flask import Flask
-app = Flask(__name__)
-from app import views
