--- app/a_model.py
@@ -1,7 +1,0 @@
-def ModelIt(fromUser  = 'Default', births = []):
- in_month = len(births)
- result = in_month
- if fromUser != 'Default':
-   return result
- else:
-   return 'check your input'
