--- app/templates/example_dbtable.html
@@ -1,59 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <meta name="description" content="">
-  <meta name="author" content="">
-  <link rel="icon" href="../../favicon.ico">
-
-  <title>You can display table data!</title>
-
-  <!-- Bootstrap core CSS -->
-  <link href="../static/css/bootstrap.min.css" rel="stylesheet">
-
-  <!-- Custom styles for this template -->
-  <link href="../static/css/bootstrap-template.css" rel="stylesheet">
-
-  <!-- HTML5 shim and Respond.js IE8 support of HTML5 elements and media queries -->
-  <!--[if lt IE 9]>
-    <script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
-    <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
-  <![endif]-->
-</head>
-
-<body>
-    <div class="container">
-    <div class="header clearfix">
-      <nav>
-        <ul class="nav nav-pills pull-left">
-          <!-- don't forget to link pages back to home-->
-          <li role="presentation"><a href="/">&larr; Back to Home</a></li>
-        </ul>
-      </nav>
-    </div>
-<br><br>
-
-  <!--we can insert a table here - note that it's taking in the birth data we passed with views.py in {{ variable }} -->
-
-  <div class="container">
-  <table class="table table-hover">
-  <tr><th>index</th><th>Attendant</th><th>birth_month</th></tr>
-  {% for birth in births %}
-  <tr><td>{{ birth['index'] }}</td><td>{{ birth['attendant']}}</td><td> {{ birth['birth_month'] }}</td></tr>
-  {% endfor %}
-  </table>
-
-  <script src="https://code.jquery.com/jquery-1.10.2.min.js"></script>
-  <script src="static/js/bootstrap.min.js"></script>
-  </div><!-- /.container -->
-
-
-  <!-- Bootstrap core JavaScript
-  ================================================== -->
-  <!-- Placed at the end of the document so the pages load faster -->
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.1/jquery.min.js"></script>
-  <script src="static/js/bootstrap.min.js"></script>
-  </body>
-</html>
